requirements.txt
pandas>=2.0.3
xarray>=2023.9.0
numpy>=1.26.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.21
kafka-python>=2.0.2
pyspark
requests>=2.31.0
boto3>=1.28.25
pyyaml>=6.0.1
fastapi>=0.104.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
